src/requirements.txt
pandas
beautifulsoup4
snowflake-connector-python
